+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Refactor `KnownBadFunctions` to improve performance and prevent regex injection.
+* Make `Priority` `Copy + Clone` and pass it by value to simplify functions that use it.
 * Improve the `usage` function.
 * Improve tests.
 * Improve documentation.